Support package version resolvment
Parses the content of a R sessionInfo()$otherPkgs output
for package name and version. These are then checked against the
Anaconda cloud API and the environment.yml is filled accordingly.

If a version can not be found on Anaconda cloud, a warning is printed
and the package will not have a dedicated version in the
environment.yml.